refactor: remove Content.content_type
This field was only really useful to represent the fact that `ImageUrl`
and `Image` were the same content type (likewise Audio). Now that those
are unified, `content.content_type == content.type` in every case, so we
can remove it.

Chunk content types are still useful as they represent the type of
content reconstructed by the chunks.